requirements.txt
numpy==1.24.3
psycopg2==2.9.9
python_bcrypt==0.3.2
torch==2.3.0
transformers==4.41.2
